Investment projects summary to use data from list endpoint
