+++ src/client/actions.js
@@ -31,10 +31,7 @@ export const CONTACTS__METADATA_LOADED = 'CONTACTS__METADATA_LOADED'
 export const ORDERS__LOADED = 'ORDERS__LOADED'
 export const ORDERS__METADATA_LOADED = 'ORDERS__METADATA_LOADED'
 
-export const DATA_SUMMARY_PICKER__RANGE_SELECTED =
-  'DATA_SUMMARY_PICKER__RANGE_SELECTED'
-export const DATA_SUMMARY_PICKER__ACCESSIBLE_TOGGLE =
-  'DATA_SUMMARY_PICKER__ACCESSIBLE_TOGGLE'
+export const DATA_SUMMARY__ACCESSIBLE_TOGGLE = 'DATA_SUMMARY__ACCESSIBLE_TOGGLE'
 
 export const EVENTS__LOADED = 'EVENTS__LOADED'
 export const EVENTS__METADATA_LOADED = 'EVENTS__METADATA_LOADED'
