+++ src/client/components/DataSummary/state.js
@@ -1,3 +1,3 @@
-export const ID = 'dataSummaryPicker'
+export const ID = 'dataSummary'
 
 export const state2props = (state) => state[ID]
