+++ src/client/components/MyInvestmentProjects/tasks.js
@@ -1,5 +1,11 @@
 import { apiProxyAxios } from '../Task/utils'
 
+const summaryToDataRange = ({ summary, adviser }) =>
+  Object.keys(summary).map((key) => ({
+    link: `/investments/projects?stage=${summary[key].id}&adviser=${adviser.id}`,
+    ...summary[key],
+  }))
+
 export const fetchMyInvestmentsList = ({
   limit = 10,
   page = 1,
@@ -13,6 +19,7 @@ export const fetchMyInvestmentsList = ({
     offset: Math.min(limit * (page - 1), 10000 - limit) || 0,
     adviser: adviser.id,
     sortby: sort,
+    show_summary: true,
   }
 
   if (filter !== 'all-stages') {
@@ -21,5 +28,8 @@ export const fetchMyInvestmentsList = ({
 
   return apiProxyAxios
     .post('/v3/search/investment_project', payload)
-    .then(({ data }) => data)
+    .then(({ data: { summary, ...rest } }) => ({
+      summary: summaryToDataRange({ summary, adviser }),
+      ...rest,
+    }))
 }
