+++ src/client/components/index.jsx
@@ -12,7 +12,7 @@ export { default as CollectionHeaderRow } from './CollectionList/CollectionHeade
 export { default as CompanyLists } from './CompanyLists'
 export { default as DeleteCompanyListSection } from './CompanyLists/DeleteCompanyListSection'
 export { default as Dashboard } from './Dashboard'
-export { default as DataSummaryPicker } from './DataSummaryPicker'
+export { default as DataSummary } from './DataSummary'
 export { default as DataTable } from './DataTable'
 export { default as DateField } from './DateField'
 export { default as EntityList } from './EntityList'
@@ -70,7 +70,7 @@ export { default as CheckboxGroupField } from './CheckboxGroupField'
 export { default as RoutedCheckboxGroupField } from './RoutedCheckboxGroupField'
 export { default as DashboardTabs } from './DashboardTabs'
 export { default as InvestmentReminders } from './InvestmentReminders'
-export { default as InvestmentProjectSummary } from './InvestmentProjectSummary'
+export { default as InvestmentProjectSummary } from './MyInvestmentProjects/InvestmentProjectSummary'
 export { default as Search } from './Search'
 export { default as Timeline } from './Timeline'
 export { default as Container } from './Layout/Container'
