+++ src/client/provider.jsx
@@ -92,8 +92,8 @@ import myInvestmentProjectsReducer from './components/MyInvestmentProjects/reduc
 import { ID as INVESTMENT_SUMMARY_ID } from './components/InvestmentProjectSummary/state'
 import investmentSummaryReducer from './components/InvestmentProjectSummary/reducer'
 
-import { ID as DATA_SUMMARY_PICKER_ID } from './components/DataSummaryPicker/state'
-import dataSummaryPickerReducer from './components/DataSummaryPicker/reducer'
+import { ID as DATA_SUMMARY_ID } from './components/DataSummary/state'
+import dataSummaryReducer from './components/DataSummary/reducer'
 
 import { ID as INVESTMENT_REMINDERS_ID } from './components/InvestmentReminders/state'
 import investmentRemindersReducer from './components/InvestmentReminders/reducer'
@@ -174,7 +174,7 @@ const store = createStore(
     [MY_INVESTMENT_PROJECTS_ID]: myInvestmentProjectsReducer,
     [CHECK_FOR_INVESTMENTS_ID]: personalDashboardReducer,
     [INVESTMENT_SUMMARY_ID]: investmentSummaryReducer,
-    [DATA_SUMMARY_PICKER_ID]: dataSummaryPickerReducer,
+    [DATA_SUMMARY_ID]: dataSummaryReducer,
     [INVESTMENT_REMINDERS_ID]: investmentRemindersReducer,
     [CONTACTS_LIST_ID]: contactsReducer,
     [COMPANY_CONTACTS_LIST_ID]: contactsReducer,
